requirements.txt
antlr4-python3-runtime==4.13.0
asgiref==3.10.0
blinker==1.9.0
click==8.3.0
Django==5.2.7
djangorestframework==3.16.1
Flask==3.1.2
gunicorn==23.0.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
sqlparse==0.5.3
Werkzeug==3.1.3